Moves DynamicTree data types into their own files.
